+++ bqsql_parser/pkg/bqsql_parser.d.ts
@@ -1,6 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-
-export function parse(bqsql: string): any;
-
-export function suggest(bqsql: string, line: number, column: number): any;